Remove data path directory each time controller starts
Fixes a bug when calling rwlock() that checks if a tag already
exists by checking if the path exists. If the path exists from a
preivous invocation of the controller, the controller will not
have created a lock for the tag yet in memory. May want to update
later to persist data even if the controller is stopped.